LayoutTests/fast/text/simple-line-layout-text-stroke-width-expected.txt

PASS internals.repaintRectsAsText().indexOf('90') is not -1
PASS successfullyParsed is true

TEST COMPLETE
foobar
